More extensive compiler optimizations.
We optimize the kernel globally for the Cortex-A55 as for Atoll CPUs, it composes the little cluster thats also abundant (6 cores out of 8). We also set the flag `-ffast-math` on crypto and lib/routines modules for further squeezing. Note that even though that flag may generate erroneous code, this is a custom kernel after all, so let give ourselves such luxury.